Solve for long numbers

+++ src/main/java/ArrayManipulation.java
@@ -1,14 +1,14 @@
 public class ArrayManipulation {
     static long arrayManipulation(int n, int[][] queries) {
-        int[] arr = new int[n];
-        int result = -1;
+        long[] arr = new long[n];
+        long result = -1;
 
         for (int i = 0; i < queries.length; i++) {
             int startIndex = queries[i][0] - 1;
             int endIndex = queries[i][1] - 1;
             int increment = queries[i][2];
 
-            int largestValue = incrementArray(arr, startIndex, endIndex, increment);
+            long largestValue = incrementArray(arr, startIndex, endIndex, increment);
             if (largestValue > result){
                 result = largestValue;
             }
@@ -17,8 +17,8 @@ public class ArrayManipulation {
         return result;
     }
 
-    private static int incrementArray(int[] arr, int startIndex, int endIndex, int increment) {
-        int result = -1;
+    private static long incrementArray(long[] arr, int startIndex, int endIndex, int increment) {
+        long result = -1;
 
         for (int i = startIndex; i <= endIndex; i++) {
             arr[i] += increment;
